Login working, tabs starting to work, plumbing and to the left crashing
I get this error:
2016-04-29 11:03:38.335 Guaranteed Pricing[2377:47661] Warning: Attempt
to present <UITabBarController: 0x7ff1929373a0> on
<Guaranteed_Pricing.Login: 0x7ff192923db0> whose view is not in the
window hierarchy!
Request returned!
2016-04-29 11:03:49.686 Guaranteed Pricing[2377:47661] *** Assertion
failure in -[UITableView
dequeueReusableCellWithIdentifier:forIndexPath:],
/BuildRoot/Library/Caches/com.apple.xbs/Sources/UIKit_Sim/UIKit-3512.30.
14/UITableView.m:6564
2016-04-29 11:03:49.705 Guaranteed Pricing[2377:47661] *** Terminating
app due to uncaught exception 'NSInternalInconsistencyException',
reason: 'unable to dequeue a cell with identifier item - must register
a nib or a class for the identifier or connect a prototype cell in a
storyboard'

when I click on plumbing or handyman. Otherwise, the app is working
well.